taobei/requirements.txt
Flask ~= 1.0.2
flask-login ~= 0.4.1
Flask-Migrate ~= 2.3.0
Flask-PyMongo ~= 2.2.0
flask-redis ~= 0.3.0
Flask-SQLAlchemy ~= 2.3.2
flask-wtf ~= 0.14.2
gevent ~= 1.3.7
Jinja2 ~= 2.10
marshmallow ~= 3.0.0b20
pillow ~= 5.3.0
python-etcd ~= 0.4.5
requests ~= 2.20.1